--- a/src/test/regress/expected/create_index_optimizer.out
+++ b/src/test/regress/expected/create_index_optimizer.out
@@ -375,16 +375,19 @@ EXPLAIN (COSTS OFF)
 SELECT * FROM fast_emp4000
     WHERE home_base @ '(200,200),(2000,1000)'::box
     ORDER BY (home_base[0])[0];
-                             QUERY PLAN                             
---------------------------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)
-   Merge Key: ((home_base[0])[0])
-   ->  Sort
-         Sort Key: ((home_base[0])[0])
-         ->  Index Scan using grect2ind on fast_emp4000
-               Index Cond: home_base @ '(2000,1000),(200,200)'::box
- Optimizer: legacy query optimizer
-(7 rows)
+                                   QUERY PLAN                                   
+--------------------------------------------------------------------------------
+ Result
+   ->  Gather Motion 3:1  (slice1; segments: 3)
+         Merge Key: ((home_base[0])[0])
+         ->  Sort
+               Sort Key: ((home_base[0])[0])
+               ->  Result
+                     ->  Index Scan using grect2ind on fast_emp4000
+                           Index Cond: home_base @ '(2000,1000),(200,200)'::box
+                           Filter: home_base @ '(2000,1000),(200,200)'::box
+ Optimizer: PQO version 2.68.0
+(10 rows)
 
 SELECT * FROM fast_emp4000
     WHERE home_base @ '(200,200),(2000,1000)'::box
@@ -404,8 +407,9 @@ SELECT count(*) FROM fast_emp4000 WHERE home_base && '(1000,1000,0,0)'::box;
          ->  Aggregate
                ->  Index Scan using grect2ind on fast_emp4000
                      Index Cond: home_base && '(1000,1000),(0,0)'::box
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: home_base && '(1000,1000),(0,0)'::box
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM fast_emp4000 WHERE home_base && '(1000,1000,0,0)'::box;
  count 
@@ -434,16 +438,19 @@ SELECT count(*) FROM fast_emp4000 WHERE home_base IS NULL;
 EXPLAIN (COSTS OFF)
 SELECT * FROM polygon_tbl WHERE f1 ~ '((1,1),(2,2),(2,1))'::polygon
     ORDER BY (poly_center(f1))[0];
-                          QUERY PLAN                           
----------------------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)
-   Merge Key: ((poly_center(f1))[0])
-   ->  Sort
-         Sort Key: ((poly_center(f1))[0])
-         ->  Index Scan using gpolygonind on polygon_tbl
-               Index Cond: f1 ~ '((1,1),(2,2),(2,1))'::polygon
- Optimizer: legacy query optimizer
-(7 rows)
+                                QUERY PLAN                                 
+---------------------------------------------------------------------------
+ Result
+   ->  Gather Motion 3:1  (slice1; segments: 3)
+         Merge Key: ((poly_center(f1))[0])
+         ->  Sort
+               Sort Key: ((poly_center(f1))[0])
+               ->  Result
+                     ->  Index Scan using gpolygonind on polygon_tbl
+                           Index Cond: f1 ~ '((1,1),(2,2),(2,1))'::polygon
+                           Filter: f1 ~ '((1,1),(2,2),(2,1))'::polygon
+ Optimizer: PQO version 2.68.0
+(10 rows)
 
 SELECT * FROM polygon_tbl WHERE f1 ~ '((1,1),(2,2),(2,1))'::polygon
     ORDER BY (poly_center(f1))[0];
@@ -455,16 +462,18 @@ SELECT * FROM polygon_tbl WHERE f1 ~ '((1,1),(2,2),(2,1))'::polygon
 EXPLAIN (COSTS OFF)
 SELECT * FROM circle_tbl WHERE f1 && circle(point(1,-2), 1)
     ORDER BY area(f1);
-                      QUERY PLAN                       
--------------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)
-   Merge Key: (area(f1))
+                            QUERY PLAN                             
+-------------------------------------------------------------------
+ Result
    ->  Sort
          Sort Key: (area(f1))
-         ->  Index Scan using gcircleind on circle_tbl
-               Index Cond: f1 && '<(1,-2),1>'::circle
- Optimizer: legacy query optimizer
-(7 rows)
+         ->  Result
+               ->  Gather Motion 3:1  (slice1; segments: 3)
+                     ->  Index Scan using gcircleind on circle_tbl
+                           Index Cond: f1 && '<(1,-2),1>'::circle
+                           Filter: f1 && '<(1,-2),1>'::circle
+ Optimizer: PQO version 2.68.0
+(9 rows)
 
 SELECT * FROM circle_tbl WHERE f1 && circle(point(1,-2), 1)
     ORDER BY area(f1);
@@ -485,8 +494,9 @@ SELECT count(*) FROM gpolygon_tbl WHERE f1 && '(1000,1000,0,0)'::polygon;
          ->  Aggregate
                ->  Index Scan using ggpolygonind on gpolygon_tbl
                      Index Cond: f1 && '((1000,1000),(0,0))'::polygon
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 && '((1000,1000),(0,0))'::polygon
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM gpolygon_tbl WHERE f1 && '(1000,1000,0,0)'::polygon;
  count 
@@ -503,8 +513,9 @@ SELECT count(*) FROM gcircle_tbl WHERE f1 && '<(500,500),500>'::circle;
          ->  Aggregate
                ->  Index Scan using ggcircleind on gcircle_tbl
                      Index Cond: f1 && '<(500,500),500>'::circle
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 && '<(500,500),500>'::circle
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM gcircle_tbl WHERE f1 && '<(500,500),500>'::circle;
  count 
@@ -521,8 +532,9 @@ SELECT count(*) FROM point_tbl WHERE f1 <@ box '(0,0,100,100)';
          ->  Aggregate
                ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 <@ '(100,100),(0,0)'::box
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 <@ '(100,100),(0,0)'::box
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl WHERE f1 <@ box '(0,0,100,100)';
  count 
@@ -538,9 +550,10 @@ SELECT count(*) FROM point_tbl WHERE box '(0,0,100,100)' @> f1;
    ->  Gather Motion 3:1  (slice1; segments: 3)
          ->  Aggregate
                ->  Index Scan using gpointind on point_tbl
-                     Index Cond: '(100,100),(0,0)'::box @> f1
- Optimizer: legacy query optimizer
-(6 rows)
+                     Index Cond: f1 <@ '(100,100),(0,0)'::box
+                     Filter: f1 <@ '(100,100),(0,0)'::box
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl WHERE box '(0,0,100,100)' @> f1;
  count 
@@ -557,8 +570,9 @@ SELECT count(*) FROM point_tbl WHERE f1 <@ polygon '(0,0),(0,100),(100,100),(50,
          ->  Aggregate
                ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 <@ '((0,0),(0,100),(100,100),(50,50),(100,0),(0,0))'::polygon
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 <@ '((0,0),(0,100),(100,100),(50,50),(100,0),(0,0))'::polygon
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl WHERE f1 <@ polygon '(0,0),(0,100),(100,100),(50,50),(100,0),(0,0)';
  count 
@@ -575,8 +589,9 @@ SELECT count(*) FROM point_tbl WHERE f1 <@ circle '<(50,50),50>';
          ->  Aggregate
                ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 <@ '<(50,50),50>'::circle
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 <@ '<(50,50),50>'::circle
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl WHERE f1 <@ circle '<(50,50),50>';
  count 
@@ -586,15 +601,16 @@ SELECT count(*) FROM point_tbl WHERE f1 <@ circle '<(50,50),50>';
 
 EXPLAIN (COSTS OFF)
 SELECT count(*) FROM point_tbl p WHERE p.f1 << '(0.0, 0.0)';
-                         QUERY PLAN                          
--------------------------------------------------------------
+                        QUERY PLAN                         
+-----------------------------------------------------------
  Aggregate
    ->  Gather Motion 3:1  (slice1; segments: 3)
          ->  Aggregate
-               ->  Index Scan using gpointind on point_tbl p
+               ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 << '(0,0)'::point
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 << '(0,0)'::point
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl p WHERE p.f1 << '(0.0, 0.0)';
  count 
@@ -604,15 +620,16 @@ SELECT count(*) FROM point_tbl p WHERE p.f1 << '(0.0, 0.0)';
 
 EXPLAIN (COSTS OFF)
 SELECT count(*) FROM point_tbl p WHERE p.f1 >> '(0.0, 0.0)';
-                         QUERY PLAN                          
--------------------------------------------------------------
+                        QUERY PLAN                         
+-----------------------------------------------------------
  Aggregate
    ->  Gather Motion 3:1  (slice1; segments: 3)
          ->  Aggregate
-               ->  Index Scan using gpointind on point_tbl p
+               ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 >> '(0,0)'::point
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 >> '(0,0)'::point
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl p WHERE p.f1 >> '(0.0, 0.0)';
  count 
@@ -622,15 +639,16 @@ SELECT count(*) FROM point_tbl p WHERE p.f1 >> '(0.0, 0.0)';
 
 EXPLAIN (COSTS OFF)
 SELECT count(*) FROM point_tbl p WHERE p.f1 <^ '(0.0, 0.0)';
-                         QUERY PLAN                          
--------------------------------------------------------------
+                        QUERY PLAN                         
+-----------------------------------------------------------
  Aggregate
    ->  Gather Motion 3:1  (slice1; segments: 3)
          ->  Aggregate
-               ->  Index Scan using gpointind on point_tbl p
+               ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 <^ '(0,0)'::point
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 <^ '(0,0)'::point
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl p WHERE p.f1 <^ '(0.0, 0.0)';
  count 
@@ -640,15 +658,16 @@ SELECT count(*) FROM point_tbl p WHERE p.f1 <^ '(0.0, 0.0)';
 
 EXPLAIN (COSTS OFF)
 SELECT count(*) FROM point_tbl p WHERE p.f1 >^ '(0.0, 0.0)';
-                         QUERY PLAN                          
--------------------------------------------------------------
+                        QUERY PLAN                         
+-----------------------------------------------------------
  Aggregate
    ->  Gather Motion 3:1  (slice1; segments: 3)
          ->  Aggregate
-               ->  Index Scan using gpointind on point_tbl p
+               ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 >^ '(0,0)'::point
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 >^ '(0,0)'::point
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl p WHERE p.f1 >^ '(0.0, 0.0)';
  count 
@@ -658,15 +677,16 @@ SELECT count(*) FROM point_tbl p WHERE p.f1 >^ '(0.0, 0.0)';
 
 EXPLAIN (COSTS OFF)
 SELECT count(*) FROM point_tbl p WHERE p.f1 ~= '(-5, -12)';
-                         QUERY PLAN                          
--------------------------------------------------------------
+                        QUERY PLAN                         
+-----------------------------------------------------------
  Aggregate
    ->  Gather Motion 3:1  (slice1; segments: 3)
          ->  Aggregate
-               ->  Index Scan using gpointind on point_tbl p
+               ->  Index Scan using gpointind on point_tbl
                      Index Cond: f1 ~= '(-5,-12)'::point
- Optimizer: legacy query optimizer
-(6 rows)
+                     Filter: f1 ~= '(-5,-12)'::point
+ Optimizer: PQO version 2.68.0
+(7 rows)
 
 SELECT count(*) FROM point_tbl p WHERE p.f1 ~= '(-5, -12)';
  count 
@@ -738,15 +758,18 @@ SELECT * FROM point_tbl WHERE f1 IS NOT NULL ORDER BY f1 <-> '0,1';
 
 EXPLAIN (COSTS OFF)
 SELECT * FROM point_tbl WHERE f1 <@ '(-10,-10),(10,10)':: box ORDER BY f1 <-> '0,1';
-                     QUERY PLAN                     
-----------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)
-   Merge Key: (f1 <-> '(0,1)'::point)
-   ->  Index Scan using gpointind on point_tbl
-         Index Cond: f1 <@ '(10,10),(-10,-10)'::box
-         Order By: f1 <-> '(0,1)'::point
- Optimizer: legacy query optimizer
-(6 rows)
+                              QUERY PLAN                              
+----------------------------------------------------------------------
+ Result
+   ->  Sort
+         Sort Key: (f1 <-> '(0,1)'::point)
+         ->  Result
+               ->  Gather Motion 3:1  (slice1; segments: 3)
+                     ->  Index Scan using gpointind on point_tbl
+                           Index Cond: f1 <@ '(10,10),(-10,-10)'::box
+                           Filter: f1 <@ '(10,10),(-10,-10)'::box
+ Optimizer: PQO version 2.68.0
+(9 rows)
 
 SELECT * FROM point_tbl WHERE f1 <@ '(-10,-10),(10,10)':: box ORDER BY f1 <-> '0,1';
    f1    
@@ -1302,17 +1325,17 @@ SET enable_indexscan = OFF;
 SET enable_bitmapscan = ON;
 EXPLAIN (COSTS OFF)
 SELECT * FROM point_tbl WHERE f1 <@ '(-10,-10),(10,10)':: box ORDER BY f1 <-> '0,1';
-                           QUERY PLAN                           
-----------------------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)
-   Merge Key: (f1 <-> '(0,1)'::point)
+                              QUERY PLAN                              
+----------------------------------------------------------------------
+ Result
    ->  Sort
          Sort Key: (f1 <-> '(0,1)'::point)
-         ->  Bitmap Heap Scan on point_tbl
-               Recheck Cond: f1 <@ '(10,10),(-10,-10)'::box
-               ->  Bitmap Index Scan on gpointind
-                     Index Cond: f1 <@ '(10,10),(-10,-10)'::box
- Optimizer: legacy query optimizer
+         ->  Result
+               ->  Gather Motion 3:1  (slice1; segments: 3)
+                     ->  Index Scan using gpointind on point_tbl
+                           Index Cond: f1 <@ '(10,10),(-10,-10)'::box
+                           Filter: f1 <@ '(10,10),(-10,-10)'::box
+ Optimizer: PQO version 2.68.0
 (9 rows)
 
 SELECT * FROM point_tbl WHERE f1 <@ '(-10,-10),(10,10)':: box ORDER BY f1 <-> '0,1';
@@ -2488,6 +2511,8 @@ ERROR:  hash indexes are not supported
 -- Test functional index
 --
 CREATE TABLE func_index_heap (f1 text, f2 text);
+NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named 'f1' as the Greenplum Database data distribution key for this table.
+HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
 CREATE INDEX func_index_index on func_index_heap (textcat(f1,f2));
 INSERT INTO func_index_heap VALUES('ABC','DEF');
 INSERT INTO func_index_heap VALUES('AB','CDEFG');
@@ -2501,6 +2526,8 @@ INSERT INTO func_index_heap VALUES('QWERTY');
 --
 DROP TABLE func_index_heap;
 CREATE TABLE func_index_heap (f1 text, f2 text);
+NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named 'f1' as the Greenplum Database data distribution key for this table.
+HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
 CREATE  INDEX func_index_index on func_index_heap ((f1 || f2) text_ops);
 INSERT INTO func_index_heap VALUES('ABC','DEF');
 INSERT INTO func_index_heap VALUES('AB','CDEFG');
@@ -2952,6 +2979,9 @@ SELECT count(*) FROM tenk1
 --
 CREATE TABLE dupindexcols AS
   SELECT unique1 as id, stringu2::text as f1 FROM tenk1;
+NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause. Creating a NULL policy entry.
+NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column(s) named 'id' as the Greenplum Database data distribution key for this table.
+HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
 CREATE INDEX dupindexcols_i ON dupindexcols (f1, id, f1 text_pattern_ops);
 ANALYZE dupindexcols;
 EXPLAIN (COSTS OFF)