--- a/src/test/regress/expected/qp_gist_indexes3.out
+++ b/src/test/regress/expected/qp_gist_indexes3.out
@@ -134,6 +134,7 @@ SELECT id, property AS "ProperTee" FROM GistTable3
 ------------------------------------------------------------------------------
 -- Encourage the optimizer to use indexes rather than sequential table scans.
 SET enable_seqscan=False;
+SET optimizer_enable_tablescan=False;
 -- Note that "=" for geometric data types means equal AREA, NOT COORDINATES.
 -- The "~=" operator means that the coordinate values, not just the area,
 -- are the same.
@@ -144,7 +145,6 @@ SELECT id, property AS "Property" FROM GistTable3
  999 | (999,999),(998,998)
 (1 row)
 
--- start_ignore
 EXPLAIN SELECT id, property AS "Property" FROM GistTable3
  WHERE property ~= '( (999,999), (998,998) )';
                                     QUERY PLAN                                    
@@ -158,7 +158,6 @@ EXPLAIN SELECT id, property AS "Property" FROM GistTable3
  Optimizer status: legacy query optimizer
 (7 rows)
 
--- end_ignore
 VACUUM ANALYZE GistTable3;
 SELECT id, property AS "ProperTee" FROM GistTable3
  WHERE property ~= '( (999,999), (998,998) )';
@@ -167,7 +166,6 @@ SELECT id, property AS "ProperTee" FROM GistTable3
  999 | (999,999),(998,998)
 (1 row)
 
--- start_ignore
 EXPLAIN SELECT id, property AS "ProperTee" FROM GistTable3
  WHERE property ~= '( (999,999), (998,998) )';
                                       QUERY PLAN                                       
@@ -179,7 +177,6 @@ EXPLAIN SELECT id, property AS "ProperTee" FROM GistTable3
  Optimizer status: legacy query optimizer
 (5 rows)
 
--- end_ignore
 -- ----------------------------------------------------------------------
 -- Test: test07Reindex.sql
 -- ----------------------------------------------------------------------
@@ -195,6 +192,7 @@ EXPLAIN SELECT id, property AS "ProperTee" FROM GistTable3
 ------------------------------------------------------------------------------
 -- Encourage the optimizer to use indexes rather than sequential table scans.
 SET enable_seqscan=False;
+SET optimizer_enable_tablescan=False;
 REINDEX INDEX GistIndex3a;
 REINDEX TABLE GistTable3;
 -- Note that "=" for geometric data types means equal AREA, NOT COORDINATES.
@@ -207,7 +205,6 @@ SELECT id, property AS "Property" FROM GistTable3
  999 | (999,999),(998,998)
 (1 row)
 
--- start_ignore
 EXPLAIN SELECT id, property AS "Property" FROM GistTable3
  WHERE property ~= '( (999,999), (998,998) )';
                                       QUERY PLAN                                       
@@ -219,7 +216,6 @@ EXPLAIN SELECT id, property AS "Property" FROM GistTable3
  Optimizer status: legacy query optimizer
 (5 rows)
 
--- end_ignore
 -- ----------------------------------------------------------------------
 -- Test: test08UniqueAndPKey.sql
 -- ----------------------------------------------------------------------