--- a/src/test/regress/init_file
+++ b/src/test/regress/init_file
@@ -13,6 +13,7 @@ m/^WARNING:  gpmon:.*Connection refused.*/
 
 m/^ Optimizer status:.*/
 m/^ Optimizer: PQO version .*/
+m/^ Settings:.*/
 
 # There are a number of NOTICE and HINT messages around table distribution,
 # for example to inform the user that the database will pick a particular